LPD-81886 Unrelated SF

+++ modules/apps/object/object-test/src/testIntegration/java/com/liferay/object/service/test/ObjectEntryLocalServiceTest.java
@@ -7953,7 +7953,7 @@ public class ObjectEntryLocalServiceTest {
 
 			resultSet.next();
 
-			return resultSet.getLong("count");
+			return resultSet.getInt("count");
 		}
 	}
